Updated JSON methods in PipelineBuilderPanel.
Updated PipelineBuilderPanel::toJson to simply return the PipelineCreator::getPipelineJson object, and PipelineBuilderPanel::setData to expect this JSON.
Added a PipelineBuilderPanel::getPublishJson that converts the PipelineCreator JSON to a format expected by the publishWorkflow service call.
Updated the PipelineDNDListener::dragDrop method to pass the template ID of the app into the PipelineCreator.
Updated PipelineEditorPanel to use the unconverted JSON from PipelineBuilderPanel for its toJson and configure methods (for saving and restoring state).
Addded PipelineEditorPanel::getPublishJson that builds all the required JSON for the publishWorkflow service call.