Add mask over PipelineBuilderPanel for valid Apps.
Updated PipelineBuilderPanel so that it becomes masked when an App that can be appended to the Pipeline is dragged over the panel.
This give the user more feedback that the App can't be inserted into the Pipeline and only appended to its end.
Updated the DNDListeners to handle drag start, enter, leave, and cancel events and not move events.