Implemented publish action for PipelineViewPresenter.
Added AutoBean models and a factory for generating the Pipeline JSON for the service.
Added a getPublishJson method in PipelineAutoBeanUtil for converting a Pipeline to a JSON string for publishing.
Refactored the getStepName methods from PipelineViewPresenter into PipelineAutoBeanUtil.
Refactored PipelineAutoBeanUtil into an instanced class since it's used only by PipelineViewPresenter.