CORE-4094 Updates for parsing pipeline IDs on save.
Added a ServiceSaveResponse AutoBean for save service call responses.
Added a PipelineAutoBeanUtil::parseServiceSaveResponseId method to parse the Pipeline's ID from this response.
Updated PipelineViewPresenter to use this method to set the ID in the Pipeline so the user may continue editing and save the Pipeline again.
Updated the "Publish" button to a "Save" button, since the pipeline editor window remains open after saving now.
Added an icon to the "Switch View" button.